Remove twilights that form subsequent outliers in time of noon and midnight.

This function performs the following steps:

Calculate noon/midnight based on twilight.

make noon/midnight circular.

split data set into day and night

loess predictions of noon/midnight (after excluding twilights with extreme variation from predictions)

fill data gaps in predictions with baytrends::fillMissing()

find the minimum difference between a noon/midnight directly calculated from twilights and any predicted noon/midnight 4 days ahead or after candidate noon/midnight.

threshold for removing twilight is when difference between noon/midnight vs prediction overreach 0.4h, but under under 12h

Usage

noonfilter(df, show_plot)

Arguments

df

Input data with twilight times in GMT shown as 'tfirst' and 'tsecond' (yyyy-mm-dd hh:mm:ss UTC) and 'type', with being 1 (tfirst is sunrise) or 2 (tfirst is sunset).

show_plot

if TRUE, plot is produced with timing of noon/midnight and affected twilights highlighted

Value

A filtered data.frame in the same format.